LPS-64528 Rename
